feat: replace initAuthority with activateChildChain
We want to store the authority address on deployment of contract to avoid
frontrunning the initAuthority method. Meanwhile, child-chain relies on
nonce to protect re-org. Thus we have this new 'activateChildChain' function
to be able to increase the nonce and block any block deposit before it is activated.

see: https://github.com/omisego/plasma-contracts/issues/304#issuecomment-538285330